+++ src/migrations/M190904193339Shopify.php
@@ -73,7 +73,6 @@ class M190904193339Shopify extends Migration
             'barcode' => $this->string()->null(),
             'sku' => $this->string()->null(),
             'is_taxable' => $this->boolean()->defaultValue(false)->notNull(),
-            'grams' => $this->decimal(10, 2)->unsigned()->null(),
             'weight' => $this->string()->null(),
             'weight_unit' => $this->string(2)->null(),
             'inventory_tracked' => $this->boolean()->unsigned()->notNull()->defaultValue(false),