+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Changed the Shopify Admin API to use the GraphQL API
 - Changed the primary key for `ProductImage` to include the `product_id`
 - Changed `Product::$options` to a flat array
+- Removed `Product::$grams`
 - Replaced `ProductVariant::$inventory_management` with `ProductVariant::$inventory_tracked`
 
 ## 2.1.12 (Mar 31, 2025)